tryon/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
pillow